Create Stdin and Stdout
I learned this from a Hackerrank challenge. https://www.hackerrank.com/challenges/java-stdin-stdout
If you were to run this code then the output will have the string empty. This is because the nextLine will read the enter as an input after you enter the double input. To fix this you just include scan.nextLine(); to solve this problem.